Make the #db{} record private
Fixing all of the compiler errors will suss out all uses of the record
outside of couch_db and couch_db_updater. This will require adding new
APIs and so on. Once this is done we'll reorder commits so that removing
the usage and adding APIs is done before removing the definition so that
we don't break the build in public.