Remove count_changes_since/2
This is nearly impossible to do without requiring a specific type of
tree index. So instead of forcing that on all storage engines we're
going to redefine how progress is measured. Instead of being exactly the
number of changes in terms of docs changed, we'll instead just do a
difference between the current database update sequence and the
previously indexed update sequence. As we process entries we'll then
just use the update sequence of each change encountered to measure our
progress.

The downside to this is that depending on update patterns progress
meters may lie a bit. Theoretically this shouldn't be too much of an
issue except for pathological cases.