HACK: For performance benchmarking only
This is just cause its easier for us to run benchmarks repeatedly while
we can change the config between runs. Otherwise we'd have to create two
releases and swap back and forth between them. Dynamic dispatch may have
a bit of a cost here but theoretically it should be a constant cost
between the two implementations.