Remove couch_lru and use a random eviction cache
This is a bit of a Jedi Mind Trick. But if we're trying to close the
least recently used idle database we really don't care which idle
database we closed. Since we're only ever closing from pressure against
max_dbs_open, any of the idle databases are likely to be least recently
used (or else they wouldn't be idle). So, just pick one at random and
close it.

This also adds a new couch_server_monitor so that we have better
tracking of which databases are idle by monitoring every client to open
the database.

Hopefully this works.