Replace LRU cache with random eviction
This is a bit of Jedi Mind Trick. Our use of the LRU with the idleness
requirement means that often times under load we were searching through
all open databases looking for one that was idle. Given that we only hit
this sort of behavior when we're pushing against max_dbs_open, the
likelihood that an idle db is the least recently used increases (because
it wouldn't be idle otherwise). So rather than pay the tax for keeping
an LRU we just randomly evict any idle database.

Idleness is now tracked by couch_server_monitor processes that monitor
every client that opens a database. When it detects its ref count has
hit zero it tells couch_server that it is idle. couch_server verifies
idleness (because an incref message could have been in flight) and on
verified idleness adds it to the ?IDLE table. Thus when closing we just
take the first key in the table and close it. If the table is empty,
then all databases are active. Anytime we incref a db we remove its
entry from the ?IDLE table so we don't attempt to close it.

COUCHDB-3191