Replace monitor per db with monitor per client
The monitor per db apprach was a bit of a failure with the increased
message passing load on couch_server. This changes the monitor to be per
client. It works by spawning a small process that immediately increments
a counter for the db and then hibernates. On wakeup it decrements the
counter and exits.