Replace couch_lru with a random eviction cache
When couch_server breaks its usually because it can't keep up with the
messaging rates flowing into it. After some profiling and various
experiments it became clear that updating and traversing the LRU were
the main cause of slowness. This change makes an optimization to the LRU
by completely removing it. It adds a monitor process per open database
which tracks clients accessing the database. These monitors are stored
in an ets table by couch_server. The random eviction is performed by
calling `ets:first/1` on the table (so "random" here means, ets'
internal set ordering).

This greatly improves the performance of couch_server.

COUCHDB-3191