Cluster purge improvements
1. Add purge UUID tree.
- Have two purge trees:
    purge_tree: purge_seq -> UUID
    upurge_tree: UUID -> {Id, [PurgedRevs]}
- Modify fold_purged_docs to purge over two trees.
    fold_purged_docs(St, StartSeq0, UserFun, UserAcc, Options), where
    UserFun :: fun(PurgeSeq, UUID, ID, Revs)
- Add open_purged_docs(Db, UUIDs)
    given UUIDs returns [{Id, [Rev]}] looked up from upurge_tree
    Used in mem3_rep:pull_purges_from_target and
    couch_db:purge_docs with replicated_changes
    to check if purge requests with these UUIDs already exist

2. Remove hard limit on purged revisions.
- Remove check that N of purge requests in purge_trees always less
    than purged_docs_limit
- Remove "purged_docs_limit" for interactive requests. Accept as
    many requests all http allows
- Remove error purge_during_compaction_exceeded limit.
    this will not happen any more, as we only trim purge trees
    during compaction
- Modify couch_db_updater:handle_call({purge_docs, ...)
    and couch_bt_engine:purge_docs_revs to allow purge multiple IDs
    at the same time

3. Add commpaction of purge trees during compaction
Purge_trees are compacted during compaction, if number of
purges exceeded purged_docs_limit. For this, all _local/purge docs
are checked to find MinClientPSeq - min checkpointed purge_seq.
Oldest purge requests before MinClientPSeq are removed from
purge_trees
during compaction.
    - couch_db_updater has the logic that's checking for max purge seq
    before which purges can be removed from purge trees:
    handle_call(get_disposable_purge_seq, _From, Db)
    - the purge copy during compaction is happened in the compactor
    code itself. Then during finish compaction, just re-run and
    re-process any changes if purge_seq /= new_purge_seq.

4. Fix Read Repair
During couch_db:update_docs with replicated_changes mode, we first
make a check if the docs to be updated have not been purged before. We
receive "replies_by_node". We check _local/purge docs for
corresponding nodes to see if the purge_seq is up to date. If not, the update
request from this node is ignored. If the purge_seq is less than a
configurable limit out of sync, the updater sequentially scans
the purge_seq tree looking for purge requests for the given revision and
if not found can continue with the write. A checking if a doc was purged of the
read-repair is checked in fabric.

5. Modify couch_db:fold_purged_docs function to make it the order of
the oldest purge (increasing purge seq)

6. Add utc_string method into couch_util that outputs current UTC
time in a string format

7. Add tests
- test_engine_compaction test that during compaction oldest
    purge_requests are removed up to purged_docs_limit

BugzID: 68275 & 81125