disaster recovery
Place plaintext markers before each encrypted structure with all
non-secret data used in encryption in order to facility disaster
recovery.

Should files be deleted by accident and only the underlying blocks are
recovered from the device, these markers will aid in recovery
(assuming the MasterKey can be retrieved, of course).

Each file-level key is prefixed with the ascii sequence KEY followed
by a variable-length encoding of the position of the data relative to
the file, followed by the RFC 3394 wrapped key itself (encrypted with
the MasterKey).

Each binary is prefixed with the ascii sequence BIN followed by the
length of the binary (32 bits), followed by the HMAC-SHA256 over the
length and the ciphertext block (32 bytes), followed by the ciphertext
itself (of previously indicated length).

It will not be possible to determine which key decrypts any given
binary so brute-force will be necessary, the HMAC's can be used as a
very strong indication that the right combination has been found.